+++ backend/uniwiki/src/main/java/com/kiwi/uniwiki/domain/document/repository/CategoryRepository.java
@@ -3,5 +3,5 @@ package com.kiwi.uniwiki.domain.document.repository;
 import com.kiwi.uniwiki.domain.document.entity.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CategoryRepository extends JpaRepository<Category, Integer> {
+public interface CategoryRepository extends JpaRepository<Category, Short> {
 }
